# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Роботоподбор API
  version: 0.1.0
  description: >
    Стаб контракта для фронтенда. Бизнес-логика расчёта живёт на бэкенде,
    фронт только отображает готовые значения.
  license:
    name: Proprietary
servers:
  - url: /api
security:
  - sessionCookie: []
paths:
...
  /projects:
    get:
      operationId: listProjects
      summary: Проекты пользователя
      description: Только для авторизованного пользователя. Без limit возвращается весь список.
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0 }
      responses:
        '200':
          description: Расчёты пользователя
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Project' }
        '401':
          description: Требуется вход
    post:
      operationId: createProject
      summary: Создать проект
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectInput' }
      responses:
        '201':
          description: Проект создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectDetail' }
        '401':
          description: Требуется вход
...
  /admin/robot-catalog:
    post:
      operationId: importRobotCatalog
      summary: Загрузить новую версию каталога роботов
      description: >
        Файлы папки каталога роботов (структура v2: index.json, categories.json, vendors.json,
        codes.json, sources.json, data/*.json) — каждый файл в поле files, имя файла с путём
        внутри папки. Каталог собирается по тем же правилам, что и при первом запуске,
        с дополнениями организатора. Характеристики обновляются из каталога, кроме полей,
        которые правил администратор. Решения, которых нет в новой версии, не удаляются.
        С dryRun=1 возвращает отчёт без записи (ТЗ 3.3.2, 3.3.6).
      parameters:
        - { name: dryRun, in: query, required: false, schema: { type: string, enum: ['1'] } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files: { type: array, items: { type: string, format: binary } }
      responses:
        '200':
          description: Отчёт о версии каталога
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RobotCatalogImportReport' }
        '401': { description: Требуется вход }
        '403': { description: Недостаточно прав }
        '422': { description: Нет файлов, файл не читается или структура каталога не та }
...
  /admin/sources:
    post:
      operationId: createSource
      summary: Добавить источник
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DataSourceInput' }
      responses:
        '201':
          description: Источник добавлен
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataSource' }
        '400': { description: Некорректные данные }
        '401': { description: Требуется вход }
        '403': { description: Недостаточно прав }
  /admin/sources/{sourceId}:
    put:
      operationId: updateSource
      summary: Изменить источник
      parameters:
        - name: sourceId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DataSourceInput' }
      responses:
        '200':
          description: Источник обновлён
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataSource' }
        '400': { description: Некорректные данные }
        '403': { description: Недостаточно прав }
        '404': { description: Не найден }
    delete:
      operationId: deleteSource
      summary: Удалить источник
      parameters:
        - name: sourceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Удалён }
        '403': { description: Недостаточно прав }
        '404': { description: Не найден }
  /admin/parameters/{fieldId}:
    patch:
      operationId: updateParameterField
      summary: Изменить норматив паспорта
      description: >
        Правка норматива паспорта объекта: значение по умолчанию, границы, источник,
        обязательность (ТЗ 3.2.5, 3.8.1). Значение по умолчанию проверяется по границам поля.
      parameters:
        - name: fieldId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ParameterFieldPatch' }
      responses:
        '200':
          description: Поле обновлено
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParameterField' }
        '400': { description: Некорректные данные }
        '403': { description: Недостаточно прав }
        '404': { description: Поле не найдено }
  /admin/changes:
    get:
      operationId: listChanges
      summary: Журнал изменений
      description: Журнал изменений каталога, нормативов и источников (ТЗ 3.3.6).
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: entity
          in: query
          required: false
          schema: { type: string, enum: [solution, source, parameter, catalog-import] }
      responses:
        '200':
          description: Записи журнала, новые сверху
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ChangeLogEntry' }
        '401': { description: Требуется вход }
        '403': { description: Недостаточно прав }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: httpOnly-cookie сессии, выдаётся при входе и регистрации.
  schemas:
...
    ParameterField:
      type: object
      required: [id, label, kind]
      properties:
        id: { type: string }
        label: { type: string }
        hint: { type: string }
        unit: { type: string }
        section: { type: string, description: 'Заголовок группы полей в форме ввода параметров объекта' }
        kind:
          type: string
          enum: [number, text, select]
        min: { type: number }
        max: { type: number }
        defaultValue: { type: string }
        required: { type: boolean, description: 'Обязательный параметр (true) или дополнительный.' }
        source: { type: string, description: Источник значения по умолчанию (норматива). }
        options:
          type: array
          items:
            type: object
            required: [value, label]
            properties:
              value: { type: string }
              label: { type: string }
...
    ProjectInput:
      type: object
      required: [title, objectType, parameters]
      properties:
        title: { type: string }
        objectType: { type: string }
        parameters:
          type: object
          additionalProperties: { type: string }
        processes:
          type: array
          items: { type: string }
        solutionId:
          type: string
          description: Решение каталога, выбранное на шаге подбора.
        calculation:
          $ref: '#/components/schemas/CalculationResult'
        modelVersion:
          type: string
          description: Версия расчётной модели, выдавшей результат.
...
    DataSourceInput:
      type: object
      required: [title, scope, kind, actualAt, confirmed]
      properties:
        title: { type: string }
        scope: { type: string }
        kind: { type: string, enum: [price, tariff, registry, benchmark, cases, dataset, open] }
        url: { type: string }
        actualAt: { type: string, description: Дата актуализации в формате источника }
        confirmed: { type: boolean }
    DataSource:
      allOf:
        - $ref: '#/components/schemas/DataSourceInput'
        - type: object
          required: [id, updatedAt]
          properties:
            id: { type: string }
            updatedAt: { type: string, format: date-time }
    ParameterFieldPatch:
      type: object
      properties:
        defaultValue: { type: string }
        min: { type: [number, 'null'] }
        max: { type: [number, 'null'] }
        source: { type: string }
        required: { type: boolean }
        hint: { type: string }
    ChangeLogEntry:
      type: object
      required: [id, createdAt, entity, entityId, action, summary]
      properties:
        id: { type: string }
        createdAt: { type: string, format: date-time }
        userEmail: { type: string }
        entity: { type: string, enum: [solution, source, parameter, catalog-import] }
        entityId: { type: string }
        action: { type: string, enum: [create, update, delete, import] }
        summary: { type: string }
        diff:
          type: object
          description: Изменённые поля — «было → стало».
          additionalProperties:
            type: object
            properties:
              from: {}
              to: {}
...
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          required: [objectType, parameters, scenarios]
          properties:
            objectType: { type: string }
            parameters:
              type: object
              additionalProperties: { type: string }
            processes:
              type: array
              items: { type: string }
            scenarios:
              type: array
              description: Минимум три — «как есть», «покупка», «услуга/RaaS».
              items: { $ref: '#/components/schemas/ScenarioBar' }
            calculationId: { type: string }
            solutionId: { type: string }
            dataVersion: { type: string }
            modelVersion: { type: string }
            calculatedAt: { type: string, format: date-time }
...
    SensitivityFactor:
      type: object
      required: [id, label, impact, direction]
      properties:
        id: { type: string }
        label: { type: string }
        impact:
          type: number
          description: Влияние на срок окупаемости, доля от базового значения.
        direction:
          type: string
          enum: [up, down]
        low: { type: string }
        high: { type: string }
    DataConfidence:
      type: string
      enum: [confirmed, needs-review]
    Maturity:
      type: string
      enum: [operation, piloting, rnd]
...
    ScenarioBar:
      type: object
      required: [id, title, subtitle, tco, share, delta]
      properties:
        id: { type: string }
        title: { type: string }
        subtitle: { type: string }
        tco: { type: number }
        share: { type: number }
        delta: { type: string }
        detail: { type: string }
        recommended: { type: boolean }
    CostLine:
      type: object
      required: [title, amount, share, source]
      properties:
        title: { type: string }
        amount: { type: number }
        share: { type: integer }
        source: { type: string }
        confidence: { $ref: '#/components/schemas/DataConfidence' }
    CostGroup:
      type: object
      required: [id, title, amount, share, confidence, lines]
      properties:
        id: { type: string }
        title: { type: string }
        amount: { type: number }
        share: { type: integer }
        confidence: { $ref: '#/components/schemas/DataConfidence' }
        lines:
          type: array
          items: { $ref: '#/components/schemas/CostLine' }
    CalculationResult:
      type: object
      required:
        [
          id, objectTitle, meta, payback, capex, roi, scenarios, costGroups,
          opexSaving, warning, assumptions, totalTco,
        ]
      properties:
        id: { type: string }
        objectTitle: { type: string }
        meta: { type: string }
        payback: { $ref: '#/components/schemas/Kpi' }
        capex: { $ref: '#/components/schemas/Kpi' }
        roi: { $ref: '#/components/schemas/Kpi' }
        opexSaving:
          type: object
          required: [series, percent, meta]
          properties:
            series: { type: array, items: { type: number } }
            percent: { type: string }
            meta: { type: string }
        warning: { type: string }
        assumptions: { type: array, items: { type: string } }
        scenarios:
          type: array
          items: { $ref: '#/components/schemas/ScenarioBar' }
        costGroups:
          type: array
          items: { $ref: '#/components/schemas/CostGroup' }
        totalTco: { type: number }
        dataVersion:
          type: string
          description: Версия каталога и нормативов, на которых сделан расчёт (ТЗ 3.1.5).
        modelVersion:
          type: string
          description: Версия расчётной модели.
        calculatedAt: { type: string, format: date-time }
        sensitivity:
          type: array
          description: Параметры, ранжированные по влиянию на результат.
          items: { $ref: '#/components/schemas/SensitivityFactor' }
    Kpi:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: string }
        unit: { type: string }
        note: { type: string }
        trend: { type: string, enum: [up, down, none] }
...
    RobotCatalogImportReport:
      type: object
      required: [applied, version, total, fit, filled, conflicts, added, updated, unchanged, missing, keptAdmin]
      properties:
        applied: { type: boolean, description: 'false — только отчёт (dryRun), в базу ничего не записано.' }
        version:
          type: object
          required: [schema, updated]
          properties:
            schema: { type: string }
            updated: { type: string, description: 'Дата версии каталога, ДД.ММ.ГГГГ.' }
        total: { type: integer, description: Решений в версии вместе с дополнениями организатора. }
        fit:
          type: object
          description: 'Тип объекта → число решений по основаниям применимости.'
          additionalProperties:
            type: object
            required: [declared, example, inferred]
            properties:
              declared: { type: integer }
              example: { type: integer }
              inferred: { type: integer }
        filled: { type: array, items: { type: string }, description: Поля каталога, дополненные из материалов организатора. }
        conflicts: { type: array, items: { type: string }, description: 'Расхождения каталога и организатора — для ответственного за каталог.' }
        added: { type: array, items: { $ref: '#/components/schemas/NamedSolution' } }
        updated:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/NamedSolution'
              - type: object
                required: [fields]
                properties: { fields: { type: array, items: { type: string } } }
        unchanged: { type: integer }
        missing:
          type: array
          description: Были в базе из каталога, но в этой версии их нет. Не удалены.
          items: { $ref: '#/components/schemas/NamedSolution' }
        keptAdmin:
          type: array
          description: Решения, в которых сохранены правки администратора.
          items:
            allOf:
              - $ref: '#/components/schemas/NamedSolution'
              - type: object
                required: [fields]
                properties: { fields: { type: array, items: { type: string } } }
    NamedSolution:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
...
    Project:
      type: object
      required: [id, title, meta, payback, status]
      properties:
        id: { type: string }
        title: { type: string }
        meta: { type: string }
        payback: { type: string }
        status: { $ref: '#/components/schemas/Maturity' }